+++ CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-07-19
+
+### Added
+
+- DAI pool support
+- Withdrawal fees breakdown
+
+## [1.2.0] - 2025-07-18
+
+### Added
+
+- sUSDS pool support
+- handling relayer fees processing
+
+### Changed 
+
+- changed withdrawal modal steps
+
+## [1.1.0] - 2025-07-16
+
+### Added
+
+- USDS pool support
+
+### Changed
+
+- Relayer quotation logic
 
 ## [1.2.0] - 2025-07-18